+++ customer-space/src/environments/environment.ts
@@ -10,4 +10,5 @@ export const environment = {
   appName: 'ELYKIA Espace Client',
   version: APP_VERSION,
   firebase: firebaseConfig,
+  remoteConfigEnabled: true,
 };